Add dictionary.dart and Refactor hangman.dart
Added the basis for a dictionary file, to pull data from a database.
I also added the models and screens folder, as is common practice in
flutter. I tried to use the correct JSON tags from MW's website.
More to come...
